Error checking in new commands

+++ src/Flavor.java
@@ -1,8 +1,8 @@
 public class Flavor{
 	String name;
-	int ram, disks, vcpus;// RAM is in GB
+	long ram, disks, vcpus;// RAM is in GB
 
-	Flavor(String name, int ram, int disks, int vcpus){
+	Flavor(String name, long ram, long disks, long vcpus){
 		this.name = name;
 		this.ram = ram;
 		this.disks = disks;
